added mat-snackBar and split method in header username

+++ todo-tracker-frontend/src/app/dashboard/dashboard.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { TodoDataService } from '../services/todo-data.service';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 export class Todo {
   constructor(
@@ -20,7 +21,7 @@ export class Todo {
 export class DashboardComponent {
   todos: Todo[] = [];
 
-  constructor(private todoDataService: TodoDataService){}
+  constructor(private todoDataService: TodoDataService, private _snackBar: MatSnackBar){}
 
   ngOnInit() {
     // this.todoDataService.getAllTodos().subscribe(
@@ -47,16 +48,28 @@ export class DashboardComponent {
     )
   }
 
-  deleteTodo(id:any) {
-    console.log(`Delete id ${id}`);
-    this.todoDataService.deleteTodo(id).subscribe(
-      response => {
-        console.log(response);
-        this.refreshTodos();
-      },
-      error => {
-        console.log(error);
-      }
-    )
+  deleteTodo(id: any) {
+    // Display a confirmation dialog
+    const isConfirmed = window.confirm('Are you sure you want to delete this todo?');
+  
+    // Check if the user confirmed the deletion
+    if (isConfirmed) {
+      console.log(`Delete id ${id}`);
+      this.todoDataService.deleteTodo(id).subscribe(
+        response => {
+          console.log(response);
+          this._snackBar.open('Todo deleted!', 'Success', {
+            duration: 3000,
+          });
+          this.refreshTodos();
+        },
+        error => {
+          console.log(error);
+        }
+      );
+    } else {
+      console.log('Deletion canceled by user.');
+    }
   }
+  
 }
